# perf: specialize range operations for better performances