@Eh2406 2024-03-11
time|Dev|PR|%
---|----|---|---
sudoku|44s|30s|32%
elm| 274ms | 277ms| ---
large| 11ms | 10ms| 9%
slow_135 | 19ms | 6.3ms| 66%
zuse | 311 ms | 297ms| 4.5%
hobo*| 1045s | 851s|18%