@zanieb 2024-03-12
@mpizenberg does the "real world" improvement of 12s to 3s on the `bio_embeddings` requirements in `uv` make this feel more worthwhile? Does that suggest that there is a missing niche in the benchmarks for the project (i.e. that Jacob ran)?