@mpizenberg 2024-03-12
> @mpizenberg does the "real world" improvement of 12s to 3s on the bio_embeddings 

Do you know if some of the inlinings produce most of the improvements, or if they all make small improvements that total to the big difference?