@konstin 2024-03-13
> Do you know if some of the inlinings produce most of the improvements, or if they all make small improvements that total to the big difference?

This work is entirely profiling and benchmark driven. I started off with two smaller problems, `Range::contains` being slower than needed and excessive clone/drop in `prior_cause`. After the first two commits you can see in the flamegraphs in the PR description it's almost entirely `union` and `intersection` calls, being slow due to excessive clone/drop (`cor...` in the flamegraphs). We had previously seen the slow drop/clone and put our version in an `Arc`, so all you see now are clone/drop on `Arc`s.

`satisfier` was calling intersection a lot, but it wasn't interested in the intersection itself, it just wanted to know if the terms are disjoint, so i special cased this into `is_disjoint`. The union calls were calling intersection once and complement three times, so i unrolled union into a version that clones as few as possible. I don't have numbers on hand for individual changes, but each made a noticeable difference.