@Eh2406 2024-03-13
> I was previously skeptic of this kind of inlining degrading the expressiveness of the computations.

Yesterday I went back and looked at my code from when I originally proposed this. You are right to reject that code. It was inscrutable.

> does the "real world" improvement of 12s to 3s on the `bio_embeddings` requirements in `uv` make this feel more worthwhile? Does that suggest that there is a missing niche in the benchmarks for the project (i.e. that Jacob ran)?

There is definitely a gap in the benchmarking infrastructure I have available. I can't seem to get a benchmark that has the performance characteristics of your real world data. Help on this would definitely be appreciated. The closest I have is `slow_135`, which is not nearly as slow as your original report, but has a similar shaped profile. The 66% improvement for that benchmark is nicely similar to the 75% your reporting.

> We had previously seen the slow drop/clone and put our version in an `Arc`, so all you see now are clone/drop on `Arc`s.

I should be able to run all benchmarks with version wrapped in `Arc`s. I wonder if that will give me graphs that looked more like yours.