@Eh2406 2024-03-13
I added two more small optimizations on top for your review  https://github.com/pubgrub-rs/pubgrub/tree/jf/177-review