@Eh2406 2024-03-13
time|Dev|PR|%
---|----|---|---
sudoku| 44s | 22s | -50%
elm| 274ms | 249.07 ms | -9%
large| 11ms | 9.8821 ms | -7%
slow_135 | 19ms | 6.6884 ms | -63%
zuse | 311 ms | 285.94 ms | -6%
hobo*| 985.2s | 803.6s | -18%
arc_elm| 381.19 ms | 318.25 ms | -16%
arc_large| 15.674 ms | 14.214 ms | -9%
arc_slow_135 | 36.262 ms | 14.885 ms | -59%
arc_zuse | 387.93 ms | 340.78 ms | -12%